Moved all the parcel highlighting rendering logic into the Map Renderer
