+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/MapRenderer/MapRenderer.asmdef
@@ -5,7 +5,8 @@
         "GUID:b1087c5731ff68448a0a9c625bb7e52d",
         "GUID:cc6bfabbfe87b7949aa248893f89ce37",
         "GUID:3f4c1b4f55716479ea104c4254afe172",
-        "GUID:760a1d365aad58044916992b072cf2a6"
+        "GUID:760a1d365aad58044916992b072cf2a6",
+        "GUID:6055be8ebefd69e48b49212b09b47b2f"
     ],
     "optionalUnityReferences": [],
     "includePlatforms": [],
